booking-bot/requirements.txt
aiogram>=3.13.0
aiosqlite>=0.20.0
python-dotenv>=1.0.0
python-dateutil>=2.9.0

# Примечание: для Python 3.14+ все зависимости уже установлены
# Если возникают проблемы с компиляцией, используйте:
# py -m pip install aiogram aiosqlite python-dotenv python-dateutil --only-binary :all:

